Refactor: 설정 수정

+++ src/main/java/com/grishare/domain/user/User.java
@@ -28,7 +28,7 @@ public class User {
     private String userName;    // 사용자 이름
     @CreatedDate
     private LocalDateTime createdAt; // 가입날짜
-    private Long birthDay; // 회원 생년월일 erd에 없음
+    private Integer birthDay; // 회원 생년월일 erd에 없음
     private String picture; // 프로필 이미지 -> 배포여부에 따라 -> 일단 배포는 미정 안할
     // 이미지 경로를 저장하는 변수
     private String address;
